icalendar support.
Gives a new propery calendar_item.icalendar_instance.

The .instance property will still give a vobject, and icalendar is not
listed as a requirement yet - icalendar module will be imported late,
when needed.  In version 1.0, probably this will be reversed, with
vobject being an optional dependency and icalendar being used
internally.

This commit also changes the way data is stored, the new logic ensures
data is stored authoratively either in _data, _vobject_instance or
_icalendar_instance.  So, mess up .e. with
event.vobject_instance.vevent.due.value and it will be refleted in
event.data.  It wasn't so earlier, but during the save operation, the
instance would be given priority.

Resolves https://github.com/python-caldav/caldav/issues/1 - references https://github.com/python-caldav/caldav/issues/97 https://github.com/python-caldav/caldav/issues/92 https://github.com/python-caldav/caldav/issues/3 and https://github.com/python-caldav/caldav/pull/28