Newer versions of xandikos and dulwich does not work with old python versions.
We apply a conditional version lock on xandikos and dulwich.

Alternatively we could just have a static version lock.  The
conditional version lock This is sort of a drawback for the python
caldav library isolated, as changes to xandikos and dulwich will
still be prone to break tests.  For the ecosystem as such, it's
probably nice that we catch problems with newer releases of xandikos
early.