Partial revert of 075841971b603bbe5774880ee1fa6abf778b84b0
An "import *" was removed in above commit.  Yes, "import *" is bad
practice - but since those things are reimported other places, we
can't remove it in a minor release.  First we need some kind of
deprecation notice, then we need to remove it when doing a major
release.  I aim for every major release to be backward compatible
with everything that wasn't marked as deprecated in the previous
x.0.0 major release, so 3.0 is the earliest we can remove this.