Bugfix - search could give objects of wrong classes
A search without filtering on comp-type on a calendar containing a mix of events, journals and tasks should return a mix of such.  (All the examples in the RFC includes the comp-type filter, so many servers does not support this).  There were a bug in the auto-detection of comp-type, so tasks would typically be wrapped as events or vice-versa.

Also ensured the CHANGELOG.md contains all changes since 2.0.0 was released.

Fixes https://github.com/python-caldav/caldav/issues/539

https://github.com/python-caldav/caldav/issues/540