More work on the compatibility matrix
My goal now is to get a master branch with test code that will work on all the test servers I have available - and then make a release with the latest bugfixes, even if the compatibility matrix is split in an old and new edition.

Making a more or less complete compatibility matrix and server-checker seems to be A LOT of work.  Perhaps I should just accept that I will not have it done this year.

Ref https://github.com/python-caldav/caldav/issues/532

https://github.com/python-caldav/caldav/pull/549