accept icalendar event/todo objects and not only calendar
Users of the library are trying to push icalendar Event objects into the caldav library, and it fails.  The CalDAV protocol demands that each "calendar object resource" is wrapped in an icalendar Calendar, but one of the objectives with this library is that users should not need to know many details about the CalDAV protocol.  Hence this should be possible.

Solves https://github.com/python-caldav/caldav/issues/546 

Partial fix for #https://github.com/python-caldav/caldav/issues/541

Caveat: the fix has two code paths, one for the (yet to be published) icalendar>=7 and another for icalendar<7.  However, the test code will only run on icalendar>=7 - meaning that this fix has not been verified for icalendar 6.