Major refactoring of search
The calendar.search-method took tons and tons of various parameters, and many special cases in the search method makes it desirable to make recursive calls - which again is very cumbersome when there are too many parameters to keep track of.

I decided to split out the search logic into a new class Searcher that can hold most of the search-related parameters. I also decided to make a division of CalDAV-related stuff and created a new package icalendar-searcher containing everything that is not related to CalDAV.  The new package will be useful for doing client-side filtering, both for calendar servers that does not support searches very well and for search queries that are not supported by the CalDAV standard.

I do hope that this separation makes the code flow easier to understand and more flexible to adapt and develop, as well as making client-side filtering easy when needed.

https://github.com/python-caldav/caldav/pull/563
https://github.com/python-caldav/caldav/issues/562