Improved substring search
By default, the server should do a substring search.

If one explicitly specifies that one wants to do a substring search and the server is marked up as missingthe feature search.text.substring, then the library will not send this text-match filter to the server, but rather do the filtering client-side.